requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
langchain>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.2.0
langchain-google-genai>=3.0.0
langchain-text-splitters>=0.0.1
chromadb>=0.3.28
openai>=0.27.0
pypdf>=3.8.0
python-docx>=0.8.11
python-multipart>=0.0.6
aiofiles>=23.1.0
requests>=2.31.0
typing-extensions>=4.5.0
python-dotenv>=1.0.0